[netstack3] Ignore multicast bound addr for send
When a socket's bound address is multicast, don't try to use it as the
src address for sent packets. Instead, follow traditional source address
selection.

Note, this new functionality exposed an existing problem in the netsvc
ping tests: Echo Requests sent from the netstack to a multicast address
are looped back to the sender by default. With this change, the Netstack
is able to respond to its own Echo Requests, which interferes with the
expectation of observing a reply from netsvc. The fix is to disable
MULTICAST_LOOP on the ping socket.

Bug: 360187268

Change-Id: Ibee3546e2415a72532394c5a2e36fe1478fa07c3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1102472
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>