[dl] Rename ModuleHandle & LoadModule
This CL renames ModuleHandle to RuntimeModule and LoadModule to
SessionModule.

The RuntimeModule is ultimately managed by the RuntimeDynamicLinker, as
the SessionModule will only live as long as a single LinkingSession. The
LinkingSession data structure will be introduced in a forthcoming CL and
the SessionModule will become private to that class.

Bug: 341473244
Change-Id: I93027fd1766bf10d4998b9ece4273061e6871412
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1104316
Fuchsia-Auto-Submit: Caslyn Tonelli <caslyn@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>