[ffx] Move SSH error checks to frontend. Report to user.
This takes ssh errors encountered when attempting to get a remote
control proxy and sends them to the user. Previously these errors would
be logged to the daemon log and then removed, not giving clients the
insight into what may be hanging.

Now, the error is always reported to the user, with reattempts happening
the same as before, but just on the client side. If a deemed-non-fatal
error is encountered, at which point a reconnect attempt will happen,
the error is printed for the user to see.

There are three considered-fatal SSH errors: key verification failure,
permission denied, and invalid arguments. These are all errors that
cannot be fixed no matter how many reattempts occur.

Bug: 357958365
Change-Id: Ie9adf90440a547333802ec704fa6be72501dd2c9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1101154
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Darren Chan <chandarren@google.com>