[kernel][phys] List supported test environments in qemu_boot_formats
This makes the qemu_boot_formats list element's provide a
supported_environments list.  The qemu_test_kernel() tests and
the shim tests now use this to plumb the environments list for
all the boot tests using each boot format.

This results in the phys tests for linuxboot and linuxbiboot
being run on AEMU as well as QEMU, and starts running shim tests
for linuxboot and linuxbiboot that weren't being run before.
When a crosvm_env becomes available for tests to run on, this
will trivially extend to having linuxboot64 and linuxbiboot phys
and shim tests also run on crosvm.

Bug: 335547883
Change-Id: Ie8202b26ef6d00e4d47d090245365f3cb8105b55
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1049875
Commit-Queue: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>