[assembly] Specifying 'update_epoch' without 'update_version_file' is an error
Previously, this silently did nothing, but also didn't create an update
package.  Now, providing a value for 'update_epoch', but not a value
for 'update_version_file' is an error.

The default value for 'update_epoch' is still '1', but it's moved from
being a default of the rule, to a default of how the rule invokes
the creation of the update package.

This corrects in-tree usages to not cause errors on this corner case.

Downstream repos may experience new errors with the following message:
"'update_version_file' must be supplied in order to build an update package."

Bug: b/361328909
Change-Id: Ie1f6ff6fe75cd81a73ac520d41abddb3baa13370
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1105314
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: David Iserovich <diserovich@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>
Reviewed-by: Darren Chan <chandarren@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>