[kernel][top] fix bug in error case when allocating idle threads
An error in the logic when handling an error case when allocating idle
threads would cause it to go ahead and mark all of the idle threads as
created, even if it stopped early. It's not really a condition that is
good to have and survive properly, but is still more graceful than
starting additional secondary cpus that have no idle thread.

Change-Id: I212c15b1f5c44302d42939ad18db016113a5051c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1100505
Fuchsia-Auto-Submit: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>