[windowed-stats] Simple8bRleRingBuffer with min_samples
Currently, Simple8bRleRingBuffer allocates up to a maximum number of
blocks. This ensures that the memory used has a fixed maximum. However,
as each time series might compress differently, it can lead to each
time series covering a different span of time, which makes it difficult
to view them together.

With this change, we make Simple8bRleRingBuffer guarantee some minimum
of samples. This way, we can guarantee that our data always covers some
minimum time span. This always makes writing visualization easier.

This change also affects the serialization for Simple8bRleRingBuffer.
In particular, the selectors buffer no longer wrap around so that we can
more easily grow its size when needed. As a result, its head index is
now always 0 or 1, instead of being any value less than `num_selectors`,
which is the number of selectors stored in the buffer. Additionally, the
number of bytes serialized for the selector ring buffer is now
`ceil((head_index + num_selectors) / 2)`, instead of
`ceil(num_selectors / 2)`.

In addition, the sampling intervals for Simple8bRle are also adjusted
now that the meaning of `capacity` is different. The minimum sampling
period is also adjusted so that the multiplicative factors from a finer
interval to a coarser one are more consistent. Now the coarser sampling
period is all 5x or 6x of its immediate finer sampling period (e.g. 5m
is 5x of 1m, whereas previously this was 10m to 1m, which is 10x).

Bug: 337086207
Test: Updated unit tests
Change-Id: I0ae03b028b2aa6fe50ee81e5f40c3c99663ab4ac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1092488
Fuchsia-Auto-Submit: Kiet Tran <kiettran@google.com>
Reviewed-by: Devon H. O'Dell <dhobsd@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>