[starnix][input] Adds a device name ioctl
This allows us to surface the device name to a program under
Starnix, making driver interaction with that program
easier to link to a specific device.

I wasn't quite sure what the "correct" format of the
device name was supposed to be, so I tried to invent one
that I think will not mess up IDC file discovery,
for example.

Bug: 353765537
Change-Id: I3e5f0f040b2bfc1bdf9165a9cc2180bd368c926a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1103574
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>