[gpt] Separate GPT to new rust crate
This is mainly a refactoring/reorganization to support testing GPT in
isolation, although as an added benefit the new GPT rust crate is now
portable if desired.

The rough split is that the GPT crate has business logic related to
parsing GPT metadata (and is set up in the future for having logic
related to manipulating a GPT as well), and storage-host just wraps that
and provides the "runtime" of exporting Block instances for the
partitions in a GPT device.

Bug: 339491886
Change-Id: I2ece649052cce534fbf444ea76b87fb1837617c8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1105441
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>