[ffx] Support direct connections with --core flag.
This updates the ffx behavior for handling connections and providing
FIDL protocols for developers. When passing the `--core` flag to a
command, non-daemon behavior will be put to use where possible.

For this initial pass, only `RemoteControlProxy` will be provided without
the daemon when the `--core` flag is given.

This defines a device connector trait that is implemented by a default
(ssh-only at the moment) connector that takes a target spec and resolves
it to an ssh-able address.

After completion of a command, if errors are encountered AND the
`--core` flag has been passed, the underlying ssh connection will be
inspected for errors, and they will be reported to the user to aid
in debugging.

Change-Id: I804eb78614753598be36d5f47c4dc91b8d49aac4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1097107
Commit-Queue: Andrew Davies <awdavies@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>