[ffx][build] Generate build-dir level config file
This generates the ffx build-dir level config file
as part of the build vs. being part of an fx command to run externally.
This makes it so the configuration will be the same for the same build
independent of the user's behavior or habits.

Bug: b/360197942
Change-Id: I3f3c45e672f9357e1d0e4cf0b789773cf7cc7672
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1101656
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>