[kernel][vm] Supply zeroes above content size
Adds an option for writing to a VMO to assume anything above the user
provided content size is zeroes, avoiding the need to read request
those pages. This is used by stream dispatcher when expanding the
content size on a write or append, as that range is logically zero.

Change-Id: I79d96c41bd1803eb5b56d1e6e6e4ffbb2f1753aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1104694
Reviewed-by: David Worsham <dworsham@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>