[kernel][vm] Correct partial success case in commit
It is possible for the loop in CommitRangeInternal to successfully
commit all the pages in the VmCowPages, i.e. have a successful
CommitRangeLocked call, but fail to process the whole range due to
ReplacePagesWithNonLoanedLocked failing to allocate a page. In this
case the assert at the bottom of the loop would execute and fail, since
it's checking commit_status, which was ZX_OK, but we failed to process
the whole range.

Change-Id: I4dcc9a750ca44cbbe6868ec6239d67ee2e6ab914
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1106553
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Reviewed-by: David Worsham <dworsham@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>