[netstack] Ensure buffer size fairness in benches
Inform both the TCP receive buffer size and the UDP send buffer size
when running socket benchmarks so that default stack values for those
sizes are not carrying disproportionate weight in the benchmark results.

While doing this, realized that netstack3 was not properly applying the
buffer size limits in all states, so introduced a new path for that with
a new test.

Removed the change detector from the tun socket benchmarks for the
socket sizes, it's not load bearing for correctness there and we can
get rid of some complexity.

Bug: 42085918
Change-Id: I35fa73a5e39f45fd48b8e4a06239afc8ee61187e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1100455
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>