[ffx][config] Fixing build-dir level config generation
This fixes the root cause of the rollback of using the
build generated config file.

The target to generate the file was moved to be a dependency
of a target that is built by default when building the
default configuration from `fx set product.board`.

There also was a double substitution going on when the
variable had a relative build directory.

Bug: b/360197942
Change-Id: I0b51a3baa1480811196b16010a442fb54d2dcf3a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1105703
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>