[fuchsia-lockfile] Fix flake with unfair locking
fuchsia-lockfile doesn't guarantee fair locks, so it's possible for one
or more threads to not grab the lock. This changes the library to expose
if the lock attempt timed out, and to change the test to loop again if
it times out a few times to avoid flakes.

Change-Id: I1cbe32fd4072ce77c50db8ac64af20929c77ff01
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1103612
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Steven Grady <slgrady@google.com>