[windowed-stats] Adjust sampling profiles
Revert the concept of SamplingProfile to what it was before
https://fxrev.dev/1064600, which is a vector of sampling intervals and
is not unique per BufferStrategy. Revert the concept of capacity of
SamplingInterval to it was before as well, when it denoted the
guaranteed number of samples held by a time series.

In this commit, we also change the specific parameters of the sampling
profiles. They are now:
- Balanced:
  - 120x1x10s - durability 20m
  - 120x1x1m - durability 2h
  - 120x1x5m - durability 10h
  - 120x1x30m - durability 2.5d
- Granular:
  - 720x1x10s - durability 2h
  - 600x1x1m - durability 10h
  - 720x1x5m - durability 2.5d
- HighlyGranular:
  - 720x1x10s - durability 2h
  - 3600x1x1m - durability 2.5d

Note that whereas before the granularity of windows were 10s, 1m, 10m,
and 1h, they are now 10s, 1m, 5m, and 30m. This is done to make the step
from one granularity to the next more consistent, as 10s->1m->5m->30m
6x, 5x, and 6x jumps, instead of the 6x, 10x, and 6x jumps like
previously.

Bug: 350797382
Test: Updated unit tests
Change-Id: Ifb9d9630a85863fb09c5ae8c382912e820533e5b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1101216
Reviewed-by: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Sean Olson <seanolson@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Kiet Tran <kiettran@google.com>