[starnix] Centralize file system creation
Previously, a different set of file systems were available on the
CurrentTask versus the Kernel object. Now, we create all the file
systems via the CurrentTask.

However, when we boot the kernel, we need to create the root file system
without a CurrentTask because we need the root file system in order to
create the initial task. This CL restricts the set of file systems that
can be used for the root mount to remote_bundle, remotefs, and tmpfs,
all of which can be created without a task.

Change-Id: I9a8a202fbc971345cef91f24de9056b641ed967c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1105168
Commit-Queue: Adam Barth <abarth@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>