[test-manager] Add pkg resolver protocol
This change routes the pkg resolver protocol to the test manager, so
that it can provide a hermetic package resolver implementation for
tests. This will be used by the driver test realm to implement support
for subpackaged drivers, as that requires it to use the package resolver
to query for subpackages.

Bug: 330189874
Change-Id: Ia0bf13b081399a7b016ffd45572e4b34cd4d9123
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1100698
Reviewed-by: Eric Stone <ecstone@google.com>
Reviewed-by: Dale Sather <dalesat@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>