[bazel_sdk] add rule to ensure cml files are compiled
Adds a rule which is used in the fuchsia_component macros which makes
sure we are always passing in a compiled manifest when building the
component.

Previously, we were checking the string that was passed in but this is
not always correct and is fragile. This change makes sure that the end
result is a compiled manifest.

Change-Id: I9ad75ab9f50aa87c56a8f9ae7d21b11ecfe265f6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1105166
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>