[dtr] Provide resolved component to startargs
Provide the resolved test component information in the driver test realm
start args so that it can be used to discover pkg drivers. With this
change we can support loading drivers from the test package directly as
well as its subpackages. This also removes the need to manually provide
a pkg handle, because the resolved component already contains that as
well.

There are a few things that need to be improved still:
 - Should have a CF introspection for getting own resolved component

Tests: basic-driver-test
Bug: 330189874

Change-Id: I2f5d7e3607162b67b97d272db3506719c8f00169
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1075272
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>