[fuchsia.io] Add Open3 at all supported API levels
Add Open3 to fuchsia.io at all previous supported API levels. We cannot
safely call Open3 on components built against different API levels
otherwise. This should be the last time we are required to rewrite
history as part of the ongoing io2 migration.

To reduce the scope of this change, we just provide stubs in the
VFS that return NOT_SUPPORTED so that we can update the VFS libraries
separately. This will ensure that when communicating with fuchsia.io
servers, calling Open3 does not result in the channel being closed
due to an unknown interaction. This allows us to safely fallback to
Open1 by keeping the channel in a usable state.

Most partners integrating the SDK periodically update to new versions,
and the majority of fuchsia.io servers are backed by VFS libraries.
This means that once this CL lands and all pre-built binaries are
updated to use the latest SDK, we can begin calling Open3 safely.

Bug: 348698584
Change-Id: I40bfd160f258783d4db49e7dd526e187bd882d2e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1102395
API-Review: Chris Suter <csuter@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>
API-Review: Adam Barth <abarth@google.com>