[blackout] destroy test component after test for suspend/resume
The continuous load generation is causing timing issues when it is
generating load while the storage driver is suspended. To prevent this,
destroy the test component right before we suspend, so the load
generation is stopped.

This somewhat degrades the usefulness of the test, since the load
generation won't go right up to the point of the suspend and get
interrupted like it will for the reboot scenarios, but consistency in
this case is certainly something we still want so it's not completely
useless.

Bug: b/355511475
Change-Id: I9726ffe07ec6762142307d9bf3b2641e7bad9a2a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1106873
Commit-Queue: Stephen Demos <sdemos@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>