[aml-cpu] Return an error if voltage could not be scaled.
The CPU driver calls out to the Power Driver to set the voltage.

Currently there is a bug that if the power driver returns a voltage
different than the voltage that was requested, the CPU driver simply
returns ZX_OK. It currently may not even scale the frequency and
just silently return success without accurately setting the operating
point.

There is another bug that if the frequency is lowered but the call to
lower the voltage fails, the frequency is not restored to the original
value.

I added two new test cases to aml-cpu-test.cc that demonstrate the
failure mode and then fixed aml-cpu.cc such that they pass.

Bug: b/309798380
Test: aml-cpu-test
Change-Id: I31d1171ea47ca77b9ca16a89b5d56cff46881dd1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108079
Commit-Queue: Gurjant Kalsi <gkalsi@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>