[superproject] Roll llvm-project/libc [libc] Fix file collision causing test flake (#106119)
In patch #105293 tests for vfscanf were added, meant to be identical to
the fscanf tests. Unfortunately, the author forgot to rename the target
file causing an occasional test flake where one test writes to the file
while the other is trying to read it. This patch fixes the issue by
renaming the target test file for the vfscanf test.

GitOrigin-RevId: 32e6827b71850c5508a455fe49812fb562ce1d92
Original-Revision: 4f66756562a69c236872acd2a73c1dff2417a278
Roller-URL: https://ci.chromium.org/b/8738422284013552417
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I09fad7c66d7b80148a4cfb30e42eb75f5eca9bbe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108535