[sockscripter] Set non-null hints for getaddrinfo
For some reason, when I run sockscripter locally within a netns,
commands like `udp bind 127.0.0.1:80` pick the IPv4-mapped IPv6 address
version of this instead of the IPv4 version. I first thought this was a
string parsing issue, but eventually traced it through to getaddrinfo.

Apparently getaddrinfo on glibc treats a null "hints" argument as
implying `ai_flags = AI_V4MAPPED | AI_ADDRCONFIG`, which means
getaddrinfo tries to be smart and tries to return you the right kind of
address based on whether you have IPv4 or IPv6 addresses configured on
your system. This is not what we want for sockscripter, and POSIX
actually specifies that the default for `ai_flags` should be 0, so let's
set that explicitly instead.

Change-Id: Iec8f1c496fb651b6526ebb606b2bbd0a0c69fd82
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1107954
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Evan Wildenhain <wildenhain@google.com>