[netstack3] Refactor IP Forwarding
Move the IPv4/IPv6 forwarding logic into a helper function. This will
enable a follow up CL to call the function when performing multicast
forwarding.

Regrettably, forwarding requires both the packet, and the
underlying buffer from which the packet is parsed. Moving the buffer
(such as by passing it as an argument to a function) invalidates the
reference to it in the packet. Thus, the approach in this CL is to
split forwarding up into two steps:
* Determine the forwarding action (using the parsed packet), and
* Forward the packet (used the underlying buffer).

While here, I opted to write my helper functions in an IP Generic
context, rather than the original separate IPv4/IPv6 implementations.

Bug: 353329136

Change-Id: I26f83cb105f3a4739623c9892b79f5a92be74e9d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1106255
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>