[kernel] Introduce energy model for kernel RPPM
Basic data structures for storing and accessing the energy model.

An energy model consist on a well defined set of power level and a
transition matrix defining properties for transitioning form one level
to another.

A power domain consist on a collection of CPU's that are bound to the
same energy model, and when a collective power level transition happens,
they are all affected.

A registry for enforcing invariants regarding existing power domains.

While power domains and models may be updated, this is handled through
versioning, that is, every instance of a power domain is immutable. An
update is performed by creating a new instance, and replacing the
reference to the old instances to the new one.

Following CLs will introduce the versioning mechanism, the control
interface handlers, etc.

Change-Id: I498f51cc63ae61f0054a381a47424ca2fa8f6094
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1084118
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>