[network/test] Test SO_REUSEADDR listener conflicts
A bug in this logic in Netstack3 was fixed in
Ide63b3767e2e256222d8b63e6128cbffabdaa797, but it is also useful to have
a test that runs on Linux as well to validate that we match its behavior
for this case. Netstack2 is currently expected to fail a subset of these
tests.

Bug: 361402347
Test: netstack3_streamsocket_test.cm
      netstack2_syncudp_streamsocket_test.cm
      host_x64/netstack_streamsocket_test
Change-Id: I0f7ebc358ce2535eda60241751ec44adc4e4d9b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108584
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>