[geni][uart] Fix RX and RFR (ready-to-receive) watermark config.
Previously RX and RFR registers were configured with values
that were incorrect (RX watermark was written to the RFR register)
and inconsistent with the bootloader configuration.

This caused a failure where the serial engine hung up and lost the
last 4 bytes received when the host wrote multiple bytes continuously
to the UART pins (for example, when the host pastes texts to the tty).

This change removes the RX and RFR config statements so that they will
not conflict with the pre-existing configs. The bootloader is fully
responsible for setting them prior to Fuchsia boot.

Test: geni-tests
Test: See bug.
Bug: 361345380
Change-Id: I56cc1ff8124cebb38e3e2aa897dc7a6b7c23e15f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108146
Reviewed-by: Andres Oportus <andresoportus@google.com>
Reviewed-by: Will Drewry <drewry@google.com>
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>