[starnix] Add feature flag for internal wake locks
This feature flag will be used to transition to a world where the
Starnix kernel manages wake locks internally, without interacting
directly with the power framework.

Instead the interactions with the power framework will be moved to the
Starnix runner.

This will allow the Starnix runner to suspend containers individually.

Change-Id: Ia30b3ba9113e98249f13625c1701b6e916dafe49
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1109192
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Guocheng Wei <guochengwei@google.com>