[netstack3] Remove Ethernet filtering
Remove support for filtering at the Ethernet layer. This better conforms
to Linux & Netstack2 behavior, which do not perform L2 Filtering
directly within the netstack. Rather, they rely on device drivers to
perform L2 Filtering.

In the absence of L2 filtering support in the device driver (currently
the case for Fuchsia) the packets will still be filtered appropriately
in the IP layer.

The motivation for this change is multicast forwarding: There are
multicast forwarding integration tests that rely on the Netstack
delivering multicast packets to the IP layer, despite not having
link-level multicast interest in the packet.

Change-Id: I5a601b08a6d94c8e9703260793252e400b2a5429
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1109272
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>