[blackout] catch component stop error on teardown
If the test fails when the component isn't running for whatever reason,
this command will fail, which messes with the error reporting for the
real failure. This teardown is best effort anyway so it doesn't really
matter.

Change-Id: I37e3ca8e932668db52be71c62208b2f21b4ebfcd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1109532
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: Pavan Kumar Juturu <jpavankumar@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>