[recovery] Remove unused recovery images
This removes all recovery images other than the recovery-fdr
image, which is the only one which has active tests in place.

This is the first of several CLs that are cleaning up this
file, they are being done in a stack to make the changes
more clear.

Run-All-Tests: true

Bug: b/362336812
Change-Id: Ifbc748d7d789a2bf329f777b10a3e97fc2b6a905
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1107976
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Mike Comfoltey <comfoltey@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>