[tools][refresh-bazel-compdb] Fix directory paths for in-tree bazel packages.
In-tree bazel packages may start with "internal_sdk" or start with  a
Fuchsia tree path. We need to rewrite the path and treat them as local
files for all such cases.

Previously we used a heuristic that only covered "//sdk". This change
adds "//src", "//zircon" and "//vendor" to that list. These covers
most of the cases for in-tree development. This is not an exhaustive
list and in the futurewe should look for a better solution to
distinguish all these in-tree directories.

Test: fx refresh-bazel-compdb --dir //src/devices/board/drivers/vim3-devicetree
All <lib/ddk/..> directories (located in //src/lib/ddk/... are parsed
correctly in compile_commands.json. clangd didn't report any error
when opening C++ sources in //src/devices/board/drivers/
vim3-devicetree.

Change-Id: I41648296b36051eda6aa055358f03ed2f03bdede
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108940
Reviewed-by: Chase Latta <chaselatta@google.com>
Commit-Queue: Yilong Li <liyl@google.com>