[bt][hfp] Send AudioControl connect/disconnect
When a peer connects, an AudioControl should receive a call to
connect to possibly set up audio system indicators that the peer is available
for HFP usage.  This enables the audio system to start/stop the HFP
audio without a call in progress.

Also send a disconnect when a peer is disconnected, to tear down any
active audio connection and/or make it unavailable

Bug: 341116731
Test: added new tests, fx test bt-hfp-audio-gateway-tests
Change-Id: Iecfffc5f96ef93c13922a85725686742692d3e6b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065732
Reviewed-by: Matthew Kehrt <kehrt@google.com>
Commit-Queue: Marie Janssen <jamuraa@google.com>