[syslog] Remove redundant thread koid TLS slot
We already have a thread-local storage slot for the thread koid in
logging_backend_fuchsia_globals.cc. That slot is linked into a share
library, which means it is shared for the whole process. The slot
removed in this CL is in logging_backend_fuchsia.cc, which is in a
source_set. That means this slot is duplicated in each linkage unit that
uses the source set.

Change-Id: I73fa8c241a8e8015e79c09ce27e215a12675e030
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1109348
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>