[build][cpp] Disable -Walloc-size-larger-than
This is not a particularly useful error: we don't specify a particular
value for this warning, and so it defaults to PTRDIFF_MAX, and it seems
unlikely that we'd be in a situation where we're making statically-known
allocations that could exceed this value. Moreover, it's buggy.

Bug: 362618077
Change-Id: Ic093bcde20505647a23e105060dc72a7fc072d92
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1103674
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>