[trivial-allocator] Clarify PageAllocator<Memory>::Allocation release semantics
Previously, the releasing in
PageAllocator<Memory>::Allocation::release() was accomplished by calling
the destructor of the held Memory::Capability. We anticipate a use-case
where it's natural to have Memory::Capability as the phys Allocation
class, but that frees - not releases - on destruction.

In this change we update the expected Memory interface to include a
Release() trait, and then use that in release() above to give clearer,
release semantics and a more flexible layer of indirection that enables
the use of using Allocation for a Capability.

Bug:
Change-Id: I34c0d8f2f0680c86fa806c9bbf229fb095b1fa8d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108936
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>