[abi-compat] Make @discoverable contradictions non-fatal by default
If a request in a @discoverable(client="external") protocol A contains a
client_end for a @discoverable(server="platform") protocol B then that
sort of implies that there's an external server for the that protocol.
But in practice, in a few places we pass through protocols whose servers
are in the platform through external components.

This makes the this situation a warning unless you pass
--fatal-contradictions to abi-compat.

Bug: 362970530
Change-Id: I15afb17a0293fe3d26e6ba70bc5e41a2430c9cc3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1106813
Commit-Queue: Ian McKellar <ianloic@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>