[bt][avrcp] GetItemAttributes is supported for Target role
Currently, when the remote CT sends GetItemAttributes command, we
respond with an error since it's not implemented on TG side.

According to the AVRCP v1.6.2 section 6.10.4.3, When the TG supports
GetItemAttributes, the CT shall use GetItemAttributes and not
GetElementAttribute. Since our AVRCP implementation supports the
browsing, if the browse channel is successfully established, we would
expect the remote CT to send the GetItemAttributes command instead of
GetElementAttributes. Therefore, we should handle the command
appropriately so that the media attributes are returned.

Bug: b/306416869
Change-Id: I8b769166de57182385571fe76abf7f66725dcecd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/940273
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Dayeong Lee <dayeonglee@google.com>