[windowed-stats] Correct Simple8bRleRingBuffer num blocks output
PackedU4VecDeque, which is used to hold the selectors of
Simple8bRleRingBuffer, doesn't decrement the len when the oldest
selector is requested to be evicted. This commit fixes this by
decrementing the len.

Also, the unit test case for rotating out the oldest value doesn't have
the #[test] tag on it. Here we add the test and also fix the test.

Bug: 337086207
Test: Fixed unit test
Change-Id: Ic5746af74743983ac3ad1730742a4ca3c4d0395b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1107919
Fuchsia-Auto-Submit: Kiet Tran <kiettran@google.com>
Reviewed-by: Sean Olson <seanolson@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>